Move getTsupConfig
