--- .browserslistrc
@@ -1,7 +1,0 @@
-# Browsers we support
-Chrome >= 73
-Firefox >= 78
-Edge >= 79
-Safari >= 12.0
-iOS >= 12.0
-opera >= 53
