+++ package.json
@@ -23,7 +23,7 @@
     "dev": "pnpm run watch",
     "prettier": "prettier \"{packages,examples,scripts}/**/*.{md,js,jsx,cjs,ts,tsx,json,vue}\"",
     "prettier:write": "pnpm run prettier --write",
-    "cipublish": "node scripts/publish.mjs",
+    "cipublish": "node scripts/publish.js",
     "cipublishforce": "CI=true pnpm cipublish"
   },
   "nx": {
