+++ packages/form-core/tsup.config.js
@@ -1,7 +1,7 @@
 // @ts-check
 
 import { defineConfig } from 'tsup'
-import { legacyConfig, modernConfig } from '../../getTsupConfig.js'
+import { legacyConfig, modernConfig } from '../../scripts/getTsupConfig.js'
 
 export default defineConfig([
   modernConfig({ entry: ['src/*.ts'] }),
